Move log likelihood out of test.
